Remove `slots` module, including `slot_contracts` and `slot_expressions`; switch over impl.
