+++ crates/bimm-contracts/benches/contracts.rs
@@ -1,17 +1,6 @@
-use bimm_contracts::slots::slot_contracts::{SlotDimMatcher, SlotShapeContract};
-use bimm_contracts::slots::slot_expressions::SlotDimExpr;
-use bimm_contracts::{DimExpr, DimMatcher, ShapeContract, run_periodically};
+use bimm_contracts::{ShapeContract, run_periodically};
+use bimm_contracts_macros::shape_contract;
 use criterion::{Criterion, criterion_group, criterion_main};
-
-static PATTERN: ShapeContract = ShapeContract::new(&[
-    DimMatcher::any(),
-    DimMatcher::expr(DimExpr::Param("b")),
-    DimMatcher::ellipsis(),
-    DimMatcher::expr(DimExpr::Prod(&[DimExpr::Param("h"), DimExpr::Param("p")])),
-    DimMatcher::expr(DimExpr::Prod(&[DimExpr::Param("w"), DimExpr::Param("p")])),
-    DimMatcher::expr(DimExpr::Pow(&DimExpr::Param("z"), 3)),
-    DimMatcher::expr(DimExpr::Param("c")),
-]);
 
 static BATCH: usize = 2;
 static HEIGHT: usize = 3;
@@ -20,56 +9,34 @@ static PADDING: usize = 4;
 static CHANNELS: usize = 5;
 static COLOR: usize = 4;
 
-fn bench_slot_unpack_shape(c: &mut Criterion) {
-    let shape = [
-        12,
-        BATCH,
-        1,
-        2,
-        3,
-        HEIGHT * PADDING,
-        WIDTH * PADDING,
-        COLOR,
-        CHANNELS,
-    ];
+fn bench_unpack_shape_bchpwp(c: &mut Criterion) {
+    static PATTERN: ShapeContract = shape_contract!["b", "c", "h" * "p", "w" * "p",];
 
-    static S_PATTERN: SlotShapeContract = SlotShapeContract {
-        index: &["batch", "height", "width", "padding", "color", "channels"],
-        terms: &[
-            SlotDimMatcher::any(),
-            SlotDimMatcher::expr(SlotDimExpr::Param { id: 0 }),
-            SlotDimMatcher::ellipsis(),
-            SlotDimMatcher::expr(SlotDimExpr::Prod {
-                children: &[SlotDimExpr::Param { id: 1 }, SlotDimExpr::Param { id: 3 }],
-            }),
-            SlotDimMatcher::expr(SlotDimExpr::Prod {
-                children: &[SlotDimExpr::Param { id: 2 }, SlotDimExpr::Param { id: 3 }],
-            }),
-            SlotDimMatcher::expr(SlotDimExpr::Param { id: 4 }),
-            SlotDimMatcher::expr(SlotDimExpr::Param { id: 5 }),
-        ],
-        ellipsis_pos: Some(2),
-    };
+    let shape = [BATCH, CHANNELS, HEIGHT * PADDING, WIDTH * PADDING];
+    let env = [("p", PADDING), ("c", CHANNELS)];
 
-    let bindings = [
-        None,
-        None,
-        None,
-        Some(PADDING as isize),
-        None,
-        Some(CHANNELS as isize),
-    ];
-
-    // thread_local! { static SCRATCH: RefCell<Vec<Option<isize>>> = RefCell::new(vec![None; 6]); }
-    c.bench_function("unpack_shape w/slots", |b| {
+    c.bench_function("unpack_shape hwp", |b| {
         b.iter(|| {
-            let mut env = bindings;
-            S_PATTERN._resolve(&shape, &mut env).unwrap()
+            let [b, h, w, c] = PATTERN.unpack_shape(&shape, &["b", "h", "w", "c"], &env);
+
+            assert_eq!(b, BATCH);
+            assert_eq!(h, HEIGHT);
+            assert_eq!(w, WIDTH);
+            assert_eq!(c, CHANNELS);
         })
     });
 }
-
 fn bench_unpack_shape(c: &mut Criterion) {
+    static PATTERN: ShapeContract = shape_contract![
+        _,
+        "b",
+        ...,
+        "h"*"p",
+        "w"*"p",
+        "z"^3,
+        "c"
+    ];
+
     let shape = [
         12,
         BATCH,
@@ -96,6 +63,16 @@ fn bench_unpack_shape(c: &mut Criterion) {
 }
 
 fn bench_assert_shape(c: &mut Criterion) {
+    static PATTERN: ShapeContract = shape_contract![
+        _,
+        "b",
+        ...,
+        "h"*"p",
+        "w"*"p",
+        "z"^3,
+        "c"
+    ];
+
     let shape = [
         12,
         BATCH,
@@ -117,6 +94,16 @@ fn bench_assert_shape(c: &mut Criterion) {
 }
 
 fn bench_assert_shape_every_nth(c: &mut Criterion) {
+    static PATTERN: ShapeContract = shape_contract![
+        _,
+        "b",
+        ...,
+        "h"*"p",
+        "w"*"p",
+        "z"^3,
+        "c"
+    ];
+
     let shape = [
         12,
         BATCH,
@@ -143,8 +130,8 @@ fn bench_assert_shape_every_nth(c: &mut Criterion) {
 
 criterion_group!(
     benches,
-    bench_slot_unpack_shape,
     bench_unpack_shape,
+    bench_unpack_shape_bchpwp,
     bench_assert_shape,
     bench_assert_shape_every_nth
 );
