+++ crates/bimm-contracts/benches/math.rs
@@ -1,15 +1,6 @@
 use bimm_contracts::math::maybe_iroot;
 use criterion::{Criterion, criterion_group, criterion_main};
 
-fn fake_iroot(value: isize, exp: usize) -> Option<usize> {
-    let v = (value as f64).powf(1.0 / exp as f64);
-    if v == v.round() {
-        Some(v as usize)
-    } else {
-        None
-    }
-}
-
 fn bench_maybe_iroot(c: &mut Criterion) {
     let mut idx = 2;
     c.bench_function("maybe_iroot", |b| {
@@ -19,9 +10,18 @@ fn bench_maybe_iroot(c: &mut Criterion) {
         })
     });
 
-    c.bench_function("fake_iroot", |b| {
+    fn reference(value: isize, exp: usize) -> Option<usize> {
+        let v = (value as f64).powf(1.0 / exp as f64);
+        if v == v.round() {
+            Some(v as usize)
+        } else {
+            None
+        }
+    }
+
+    c.bench_function("float reference maybe_iroot", |b| {
         b.iter(|| {
-            let _r = fake_iroot(idx, 4);
+            let _r = reference(idx, 4);
             idx += 1;
         })
     });
