+++ crates/bimm-contracts/src/expressions.rs
@@ -1,62 +1,106 @@
 //! # Dimension Expressions.
-//!
-//! A [`DimExpr`] is an algebraic constraint pattern matching a dimension.
-//! - [`DimExpr::Param`] matches a binding by name.
-//! - [`DimExpr::Negate`] matches the negation of an inner expression.
-//! - [`DimExpr::Pow`] matches an exponential power of an inner expression.
-//! - [`DimExpr::Sum`] matches the sum of a series of inner expressions.
-//! - [`DimExpr::Prod`] matches the product of a series of inner expressions.
 
-use crate::bindings::StackMap;
 use crate::math::maybe_iroot;
-use alloc::string::{String, ToString};
 use core::fmt::{Display, Formatter};
 
 /// A stack/static expression algebra for dimension sizes.
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub enum DimExpr<'a> {
     /// A parameter reference.
-    Param(&'a str),
+    Param {
+        /// The id of the parameter.
+        id: usize,
+    },
 
     /// Negation of an expression.
-    Negate(&'a DimExpr<'a>),
+    Negate {
+        /// The child expression.
+        child: &'a DimExpr<'a>,
+    },
 
     /// Exponentiation of an expression.
-    Pow(&'a DimExpr<'a>, usize),
+    Pow {
+        /// The child expression.
+        base: &'a DimExpr<'a>,
+
+        /// The exponent.
+        exp: usize,
+    },
 
     /// Sum of expressions.
-    Sum(&'a [DimExpr<'a>]),
+    Sum {
+        /// The child expressions.
+        children: &'a [DimExpr<'a>],
+    },
 
     /// Product of expressions.
-    Prod(&'a [DimExpr<'a>]),
+    Prod {
+        /// The child expressions.
+        children: &'a [DimExpr<'a>],
+    },
 }
 
-impl Display for DimExpr<'_> {
+/// Display Adapter to format `DimExprs` with a `Index`.
+pub struct ExprDisplayAdapter<'a> {
+    ///  index.
+    pub index: &'a [&'a str],
+
+    /// Expression to format.
+    pub expr: &'a DimExpr<'a>,
+}
+
+impl<'a> Display for ExprDisplayAdapter<'a> {
     fn fmt(&self, f: &mut Formatter<'_>) -> core::fmt::Result {
-        // TODO: with some lifting, we could elide more of the parentheses.
-        match self {
-            DimExpr::Param(param) => write!(f, "{param}"),
-            DimExpr::Negate(negate) => write!(f, "(-{negate})"),
-            DimExpr::Pow(base, exponent) => {
-                write!(f, "({base})^{exponent}")
-            }
-            DimExpr::Sum(values) => {
+        match self.expr {
+            DimExpr::Param { id } => write!(f, "{}", self.index[*id]),
+            DimExpr::Negate { child } => write!(
+                f,
+                "(-{})",
+                ExprDisplayAdapter {
+                    expr: child,
+                    index: self.index
+                }
+            ),
+            DimExpr::Pow { base: child, exp } => write!(
+                f,
+                "({}^{})",
+                ExprDisplayAdapter {
+                    expr: child,
+                    index: self.index
+                },
+                exp
+            ),
+            DimExpr::Sum { children } => {
                 write!(f, "(")?;
-                for (idx, expr) in values.iter().enumerate() {
+                for (idx, expr) in children.iter().enumerate() {
                     if idx > 0 {
                         write!(f, "+")?;
                     }
-                    write!(f, "{expr}")?;
+                    write!(
+                        f,
+                        "{}",
+                        ExprDisplayAdapter {
+                            expr,
+                            index: self.index
+                        }
+                    )?;
                 }
                 write!(f, ")")
             }
-            DimExpr::Prod(values) => {
+            DimExpr::Prod { children } => {
                 write!(f, "(")?;
-                for (idx, expr) in values.iter().enumerate() {
+                for (idx, expr) in children.iter().enumerate() {
                     if idx > 0 {
                         write!(f, "*")?;
                     }
-                    write!(f, "{expr}")?;
+                    write!(
+                        f,
+                        "{}",
+                        ExprDisplayAdapter {
+                            expr,
+                            index: self.index
+                        }
+                    )?;
                 }
                 write!(f, ")")
             }
@@ -65,12 +109,15 @@ impl Display for DimExpr<'_> {
 }
 
 #[derive(Debug, Clone, PartialEq, Eq)]
-enum TryEvalResult {
+enum EvalResult {
     /// The evaluated value of the expression.
-    Value(isize),
+    Value { value: isize },
 
     /// The count of unbound parameters in the expression.
-    UnboundParams(usize),
+    UnboundParams {
+        /// The count of unbound parameters.
+        count: usize,
+    },
 }
 
 /// Result of `SizeExpr::try_match()`.
@@ -81,7 +128,7 @@ enum TryEvalResult {
 /// Runtime errors (malformed expressions, too-many unbound parameters, etc.)
 /// are not represented here; and are returned as `Err(String)` from `try_match`.
 #[derive(Debug, Clone, PartialEq, Eq)]
-pub enum TryMatchResult<'a> {
+pub enum MatchResult {
     /// All params bound and expression equals target.
     Match,
 
@@ -89,7 +136,13 @@ pub enum TryMatchResult<'a> {
     Conflict,
 
     /// Expression can be solved for a single unbound param.
-    ParamConstraint(&'a str, isize),
+    ParamConstraint {
+        /// The id of the parameter.
+        id: usize,
+
+        /// The value the parameter must take to satisfy the expression.
+        value: isize,
+    },
 }
 
 impl<'a> DimExpr<'a> {
@@ -105,50 +158,48 @@ impl<'a> DimExpr<'a> {
     /// * `Value(value)` - the evaluated value of the expression.
     /// * `UnboundParams(count)` - the count of unbound parameters.
     #[must_use]
-    fn try_eval<E>(&self, env: &E) -> TryEvalResult
-    where
-        E: StackMap<'a, usize>,
-    {
+    fn try_eval(&self, env: &[Option<isize>]) -> EvalResult {
         #[inline(always)]
-        fn reduce_children<'a, B>(
+        fn reduce_children<'a>(
             exprs: &'a [DimExpr<'a>],
-            bindings: &B,
+            env: &[Option<isize>],
             zero: isize,
             op: fn(&mut isize, isize),
-        ) -> TryEvalResult
-        where
-            B: StackMap<'a, usize>,
-        {
+        ) -> EvalResult {
             let mut value = zero;
-            let mut unbound_count = 0;
+            let mut count = 0;
             for expr in exprs {
-                match expr.try_eval(bindings) {
-                    TryEvalResult::Value(v) => op(&mut value, v),
-                    TryEvalResult::UnboundParams(c) => unbound_count += c,
+                match expr.try_eval(env) {
+                    EvalResult::Value { value: v } => op(&mut value, v),
+                    EvalResult::UnboundParams { count: c } => count += c,
                 }
             }
-            if unbound_count == 0 {
-                TryEvalResult::Value(value)
+            if count == 0 {
+                EvalResult::Value { value }
             } else {
-                TryEvalResult::UnboundParams(unbound_count)
+                EvalResult::UnboundParams { count }
             }
         }
 
         match self {
-            DimExpr::Param(name) => match env.lookup(name) {
-                Some(value) => TryEvalResult::Value(value as isize),
-                None => TryEvalResult::UnboundParams(1),
+            DimExpr::Param { id } => match env[*id] {
+                Some(value) => EvalResult::Value { value },
+                None => EvalResult::UnboundParams { count: 1 },
             },
-            DimExpr::Negate(expr) => match expr.try_eval(env) {
-                TryEvalResult::Value(value) => TryEvalResult::Value(-value),
+            DimExpr::Negate { child } => match child.try_eval(env) {
+                EvalResult::Value { value } => EvalResult::Value { value: -value },
                 x => x,
             },
-            DimExpr::Pow(base, exp) => match base.try_eval(env) {
-                TryEvalResult::Value(value) => TryEvalResult::Value(value.pow(*exp as u32)),
+            DimExpr::Pow { base: child, exp } => match child.try_eval(env) {
+                EvalResult::Value { value } => EvalResult::Value {
+                    value: value.pow(*exp as u32),
+                },
                 x => x,
             },
-            DimExpr::Sum(children) => reduce_children(children, env, 0, |tmp, value| *tmp += value),
-            DimExpr::Prod(children) => {
+            DimExpr::Sum { children } => {
+                reduce_children(children, env, 0, |tmp, value| *tmp += value)
+            }
+            DimExpr::Prod { children } => {
                 reduce_children(children, env, 1, |tmp, value| *tmp *= value)
             }
         }
@@ -168,31 +219,29 @@ impl<'a> DimExpr<'a> {
     /// * `Ok(MatchResult::Constraint(name, value))` if the expression can be solved for a single unbound parameter.
     /// * `Ok(MatchResult::UnderConstrained)` if the expression cannot be solved with the current bindings.
     #[must_use]
-    pub fn try_match<E>(&'a self, target: isize, env: &E) -> Result<TryMatchResult<'a>, String>
-    where
-        E: StackMap<'a, usize>,
-    {
+    pub fn try_match(
+        &self,
+        target: isize,
+        env: &[Option<isize>],
+    ) -> Result<MatchResult, &'static str> {
         #[inline(always)]
-        fn reduce_children<'a, E>(
+        fn reduce_children<'a>(
             exprs: &'a [DimExpr<'a>],
-            env: &E,
+            env: &[Option<isize>],
             zero: isize,
             op: fn(&mut isize, isize),
-        ) -> Result<(isize, Option<&'a DimExpr<'a>>), String>
-        where
-            E: StackMap<'a, usize>,
-        {
+        ) -> Result<(isize, Option<&'a DimExpr<'a>>), &'static str> {
             let mut partial_value: isize = zero;
             let mut rem_expr = None;
             // At most one child can be unbound, and by only one parameter.
             for expr in exprs {
                 match expr.try_eval(env) {
-                    TryEvalResult::Value(value) => op(&mut partial_value, value),
-                    TryEvalResult::UnboundParams(count) => {
+                    EvalResult::Value { value } => op(&mut partial_value, value),
+                    EvalResult::UnboundParams { count } => {
                         if count == 1 && rem_expr.is_none() {
                             rem_expr = Some(expr);
                         } else {
-                            return Err("Too many unbound params".to_string());
+                            return Err("Too many unbound params.");
                         }
                     }
                 }
@@ -203,44 +252,45 @@ impl<'a> DimExpr<'a> {
         }
 
         match self {
-            DimExpr::Param(name) => {
-                if let Some(value) = env.lookup(name) {
-                    if value as isize == target {
-                        Ok(TryMatchResult::Match)
+            DimExpr::Param { id } => {
+                let id = *id;
+                if let Some(value) = env[id] {
+                    if value == target {
+                        Ok(MatchResult::Match)
                     } else {
-                        Ok(TryMatchResult::Conflict)
+                        Ok(MatchResult::Conflict)
                     }
                 } else {
-                    Ok(TryMatchResult::ParamConstraint(name, target))
+                    Ok(MatchResult::ParamConstraint { id, value: target })
                 }
             }
-            DimExpr::Negate(expr) => expr.try_match(-target, env),
-            DimExpr::Pow(base, exp) => match maybe_iroot(target, *exp) {
-                Some(root) => base.try_match(root, env),
-                None => Err("No integer solution.".to_string()),
+            DimExpr::Negate { child } => child.try_match(-target, env),
+            DimExpr::Pow { base: child, exp } => match maybe_iroot(target, *exp) {
+                Some(root) => child.try_match(root, env),
+                None => Err("No integer solution."),
             },
-            DimExpr::Sum(exprs) => {
-                let (value, rem) = reduce_children(exprs, env, 0, |acc, value| *acc += value)?;
+            DimExpr::Sum { children } => {
+                let (value, rem) = reduce_children(children, env, 0, |tmp, value| *tmp += value)?;
                 if let Some(expr) = rem {
                     expr.try_match(target - value, env)
                 } else if value == target {
-                    Ok(TryMatchResult::Match)
+                    Ok(MatchResult::Match)
                 } else {
-                    Ok(TryMatchResult::Conflict)
+                    Ok(MatchResult::Conflict)
                 }
             }
-            DimExpr::Prod(exprs) => {
-                let (value, rem) = reduce_children(exprs, env, 1, |acc, value| *acc *= value)?;
+            DimExpr::Prod { children } => {
+                let (value, rem) = reduce_children(children, env, 1, |tmp, value| *tmp *= value)?;
                 if let Some(expr) = rem {
                     if target % value != 0 {
                         // Non-integer solution
-                        return Err("No integer solution.".to_string());
+                        return Err("No integer solution.");
                     }
                     expr.try_match(target / value, env)
                 } else if value == target {
-                    Ok(TryMatchResult::Match)
+                    Ok(MatchResult::Match)
                 } else {
-                    Ok(TryMatchResult::Conflict)
+                    Ok(MatchResult::Conflict)
                 }
             }
         }
@@ -250,120 +300,167 @@ impl<'a> DimExpr<'a> {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::bindings::StackEnvironment;
+    use alloc::format;
+    use alloc::string::String;
 
     #[test]
-    fn test_try_eval() {
-        let env: StackEnvironment = &[("a", 5), ("b", 3)];
+    fn test_format() {
+        static INDEX: [&str; 5] = ["a", "b", "c", "d", "e"];
 
-        let expr = DimExpr::Param("a");
-        assert_eq!(expr.try_eval(&env), TryEvalResult::Value(5));
-        assert_eq!(expr.try_match(5, &env).unwrap(), TryMatchResult::Match);
-        assert_eq!(expr.try_match(42, &env).unwrap(), TryMatchResult::Conflict);
+        fn fmt(expr: &DimExpr) -> String {
+            format!(
+                "{}",
+                ExprDisplayAdapter {
+                    expr: &expr,
+                    index: &INDEX
+                }
+            )
+        }
 
-        let expr = DimExpr::Param("x");
-        assert_eq!(expr.try_eval(&env), TryEvalResult::UnboundParams(1));
         assert_eq!(
-            expr.try_match(5, &env).unwrap(),
-            TryMatchResult::ParamConstraint("x", 5)
+            fmt(&DimExpr::Param {
+                id: INDEX
+                    .iter()
+                    .enumerate()
+                    .find_map(|(i, k)| if *k == "a" { Some(i) } else { None })
+                    .unwrap()
+            }),
+            "a"
         );
 
-        let expr = DimExpr::Negate(&DimExpr::Param("a"));
-        assert_eq!(expr.try_eval(&env), TryEvalResult::Value(-5));
-        assert_eq!(expr.try_match(-5, &env).unwrap(), TryMatchResult::Match);
-        assert_eq!(expr.try_match(42, &env).unwrap(), TryMatchResult::Conflict);
-
-        let expr = DimExpr::Negate(&DimExpr::Param("x"));
-        assert_eq!(expr.try_eval(&env), TryEvalResult::UnboundParams(1));
-        assert_eq!(
-            expr.try_match(-5, &env).unwrap(),
-            TryMatchResult::ParamConstraint("x", 5)
-        );
-
-        let expr = DimExpr::Pow(&DimExpr::Param("a"), 3);
-        assert_eq!(expr.try_eval(&env), TryEvalResult::Value(5 * 5 * 5));
-        assert_eq!(
-            expr.try_match(5 * 5 * 5, &env).unwrap(),
-            TryMatchResult::Match
-        );
-        assert_eq!(
-            expr.try_match(4 * 4 * 4, &env).unwrap(),
-            TryMatchResult::Conflict
-        );
-
-        let expr = DimExpr::Pow(&DimExpr::Param("x"), 3);
-        assert_eq!(expr.try_eval(&env), TryEvalResult::UnboundParams(1));
-        assert_eq!(
-            expr.try_match(27, &env).unwrap(),
-            TryMatchResult::ParamConstraint("x", 3)
-        );
-
-        let expr = DimExpr::Sum(&[DimExpr::Param("a"), DimExpr::Param("b")]);
-        assert_eq!(expr.try_eval(&env), TryEvalResult::Value(8));
-        assert_eq!(expr.try_match(8, &env).unwrap(), TryMatchResult::Match);
-        assert_eq!(expr.try_match(42, &env).unwrap(), TryMatchResult::Conflict);
-
-        let expr = DimExpr::Sum(&[DimExpr::Param("x"), DimExpr::Param("b")]);
-        assert_eq!(expr.try_eval(&env), TryEvalResult::UnboundParams(1));
-        assert_eq!(
-            expr.try_match(8, &env).unwrap(),
-            TryMatchResult::ParamConstraint("x", 5)
-        );
-
-        let expr = DimExpr::Prod(&[DimExpr::Param("a"), DimExpr::Param("b")]);
-        assert_eq!(expr.try_eval(&env), TryEvalResult::Value(15));
-        assert_eq!(expr.try_match(15, &env).unwrap(), TryMatchResult::Match);
-        assert_eq!(expr.try_match(42, &env).unwrap(), TryMatchResult::Conflict);
-
-        let expr = DimExpr::Prod(&[DimExpr::Param("x"), DimExpr::Param("b")]);
-        assert_eq!(expr.try_eval(&env), TryEvalResult::UnboundParams(1));
-        assert_eq!(
-            expr.try_match(15, &env).unwrap(),
-            TryMatchResult::ParamConstraint("x", 5)
-        );
-
-        // Fancy
-        let expr = DimExpr::Sum(&[
-            DimExpr::Prod(&[DimExpr::Param("a"), DimExpr::Param("b")]),
-            DimExpr::Negate(&DimExpr::Pow(&DimExpr::Param("a"), 2)),
-        ]);
-        let target = 5 * 3 - 5 * 5; // 15 - 25 = -10
-        assert_eq!(expr.try_eval(&env), TryEvalResult::Value(target));
-        assert_eq!(expr.try_match(target, &env).unwrap(), TryMatchResult::Match);
-        assert_eq!(expr.try_match(42, &env).unwrap(), TryMatchResult::Conflict);
+        let _expr = DimExpr::Prod {
+            children: &[
+                DimExpr::Param { id: 0 },
+                DimExpr::Param { id: 1 },
+                DimExpr::Sum {
+                    children: &[
+                        DimExpr::Param { id: 2 },
+                        DimExpr::Pow {
+                            base: &DimExpr::Param { id: 3 },
+                            exp: 2,
+                        },
+                        DimExpr::Negate {
+                            child: &DimExpr::Param { id: 4 },
+                        },
+                    ],
+                },
+            ],
+        };
+        assert_eq!(fmt(&_expr), "(a*b*(c+(d^2)+(-e)))");
     }
 
     #[test]
-    fn test_too_many_unbound_params() {
-        let env: StackEnvironment = &[("a", 5), ("b", 3)];
+    fn test_eval_param() {
+        let env = [Some(5), None];
 
-        let expr = DimExpr::Sum(&[
-            DimExpr::Param("a"),
-            DimExpr::Param("b"),
-            DimExpr::Param("x"),
-            DimExpr::Param("y"),
-        ]);
-        assert_eq!(expr.try_eval(&env), TryEvalResult::UnboundParams(2));
-        assert!(expr.try_match(8, &env).is_err());
-        assert!(expr.try_match(42, &env).is_err());
+        let expr = DimExpr::Param { id: 0 };
+        assert_eq!(expr.try_eval(&env), EvalResult::Value { value: 5 });
+        assert_eq!(expr.try_match(5, &env), Ok(MatchResult::Match));
+        assert_eq!(expr.try_match(42, &env), Ok(MatchResult::Conflict));
+
+        let expr = DimExpr::Param { id: 1 };
+        assert_eq!(expr.try_eval(&env), EvalResult::UnboundParams { count: 1 });
+        assert_eq!(
+            expr.try_match(5, &env),
+            Ok(MatchResult::ParamConstraint { id: 1, value: 5 })
+        );
     }
 
     #[test]
-    fn test_pow_no_integer_solution() {
-        let env: StackEnvironment = &[("a", 5)];
+    fn try_eval_negate() {
+        let expr = DimExpr::Negate {
+            child: &DimExpr::Param { id: 0 },
+        };
 
-        let expr = DimExpr::Pow(&DimExpr::Param("a"), 3);
-        assert_eq!(expr.try_eval(&env), TryEvalResult::Value(125));
-        assert!(expr.try_match(126, &env).is_err());
+        let env = [Some(5)];
+        assert_eq!(expr.try_eval(&env), EvalResult::Value { value: -5 });
+        assert_eq!(expr.try_match(-5, &env), Ok(MatchResult::Match));
+        assert_eq!(expr.try_match(42, &env), Ok(MatchResult::Conflict));
+
+        let env = [None];
+        assert_eq!(expr.try_eval(&env), EvalResult::UnboundParams { count: 1 });
+        assert_eq!(
+            expr.try_match(-5, &env),
+            Ok(MatchResult::ParamConstraint { id: 0, value: 5 })
+        );
     }
 
     #[test]
-    fn test_prod_no_integer_solution() {
-        let env: StackEnvironment = &[("a", 5)];
+    fn try_eval_pow() {
+        let expr = DimExpr::Pow {
+            base: &DimExpr::Param { id: 0 },
+            exp: 3,
+        };
 
-        let expr = DimExpr::Prod(&[DimExpr::Param("a"), DimExpr::Param("b")]);
-        assert_eq!(expr.try_eval(&env), TryEvalResult::UnboundParams(1));
-        assert!(expr.try_match(14, &env).is_err());
-        assert!(expr.try_match(15, &env).is_ok());
+        let env = [Some(5)];
+        assert_eq!(expr.try_eval(&env), EvalResult::Value { value: 125 });
+        assert_eq!(expr.try_match(125, &env), Ok(MatchResult::Match));
+        assert_eq!(expr.try_match(42, &env), Err("No integer solution."));
+
+        let env = [None];
+        assert_eq!(expr.try_eval(&env), EvalResult::UnboundParams { count: 1 });
+        assert_eq!(
+            expr.try_match(125, &env),
+            Ok(MatchResult::ParamConstraint { id: 0, value: 5 })
+        );
+    }
+
+    #[test]
+    fn test_eval_sum() {
+        let expr = DimExpr::Sum {
+            children: &[
+                DimExpr::Param { id: 0 },
+                DimExpr::Param { id: 1 },
+                DimExpr::Param { id: 2 },
+                DimExpr::Param { id: 3 },
+            ],
+        };
+
+        let env = [Some(2), Some(3), Some(4), Some(5)];
+        assert_eq!(expr.try_eval(&env), EvalResult::Value { value: 14 });
+        assert_eq!(expr.try_match(14, &env), Ok(MatchResult::Match));
+        assert_eq!(expr.try_match(42, &env), Ok(MatchResult::Conflict));
+
+        let env = [Some(2), Some(3), None, Some(5)];
+        assert_eq!(expr.try_eval(&env), EvalResult::UnboundParams { count: 1 });
+        assert_eq!(
+            expr.try_match(14, &env),
+            Ok(MatchResult::ParamConstraint { id: 2, value: 4 })
+        );
+
+        let env = [Some(5), Some(3), None, None];
+        assert_eq!(expr.try_eval(&env), EvalResult::UnboundParams { count: 2 });
+        assert_eq!(expr.try_match(14, &env), Err("Too many unbound params."));
+    }
+
+    #[test]
+    fn test_eval_prod() {
+        let expr = DimExpr::Prod {
+            children: &[
+                DimExpr::Param { id: 0 },
+                DimExpr::Param { id: 1 },
+                DimExpr::Param { id: 2 },
+                DimExpr::Param { id: 3 },
+            ],
+        };
+
+        let env = [Some(2), Some(3), Some(4), Some(5)];
+        assert_eq!(expr.try_eval(&env), EvalResult::Value { value: 120 });
+        assert_eq!(expr.try_match(120, &env), Ok(MatchResult::Match));
+        assert_eq!(expr.try_match(42, &env), Ok(MatchResult::Conflict));
+
+        let env = [Some(2), Some(3), None, Some(5)];
+        assert_eq!(expr.try_eval(&env), EvalResult::UnboundParams { count: 1 });
+        assert_eq!(
+            expr.try_match(120, &env),
+            Ok(MatchResult::ParamConstraint { id: 2, value: 4 })
+        );
+
+        let env = [Some(1), Some(5), None, Some(5)];
+        assert_eq!(expr.try_match(40, &env), Err("No integer solution."));
+
+        let env = [Some(5), Some(3), None, None];
+        assert_eq!(expr.try_eval(&env), EvalResult::UnboundParams { count: 2 });
+        assert_eq!(expr.try_match(120, &env), Err("Too many unbound params."));
     }
 }
