+++ crates/bimm-contracts/src/lib.rs
@@ -262,7 +262,6 @@ pub mod expressions;
 pub mod macros;
 pub mod math;
 pub mod shape_argument;
-pub mod slots;
 pub mod support;
 
 pub use bindings::StackEnvironment;
