--- crates/bimm-contracts/src/slots/mod.rs
@@ -1,3 +1,0 @@
-//! # In-Development Slot Contract Support
-pub mod slot_contracts;
-pub mod slot_expressions;
